move ssh tests from utils test into own file
